+++ src/main/java/pl/dudekjunior/familyapp/models/servicies/ChildService.java
@@ -2,12 +2,13 @@ package pl.dudekjunior.familyapp.models.servicies;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import pl.dudekjunior.familyapp.models.FamilyModel;
+import pl.dudekjunior.familyapp.models.Gender;
 import pl.dudekjunior.familyapp.models.entities.ChildEntity;
 import pl.dudekjunior.familyapp.models.entities.FatherEntity;
 import pl.dudekjunior.familyapp.models.forms.ChildForm;
 import pl.dudekjunior.familyapp.models.repositories.ChildRepository;
 
+import java.time.LocalDate;
 import java.util.List;
 
 @Service
@@ -35,10 +36,23 @@ public class ChildService {
         return childEntity;
     }
 
-    public List<ChildEntity> getChildrenByFatherId(FamilyModel family) {
-        return childRepository.findByFatherId(family.getFather().getId());
-    }
     public List<ChildEntity> getChildrenByName(String childName) {
         return childRepository.findByName(childName);
+    }
+
+    public List<ChildEntity> getChildrenBySurname(String childSurname) {
+        return childRepository.findBySurname(childSurname);
+    }
+
+    public List<ChildEntity> getChildrenByPesel(String childPesel) {
+        return childRepository.findByPesel(childPesel);
+    }
+
+    public List<ChildEntity> getChildrenByBornDate(LocalDate childBornDate) {
+        return childRepository.findByBornDate(childBornDate);
+    }
+
+    public List<ChildEntity> getChildrenByGender(Gender childGender) {
+        return childRepository.findByGender(childGender);
     }
 }
